migrate automate logic to camera_fragment
